Move nested() and unbind() to utils.py
